### 2024-08-02 81f247e
update pythtb docstring
