### 2024-08-05 #7055 @rkurchin
Great, thanks! I'll add you as a reviewer. Once we find at least one more reviewer, I will trigger the start of the review issue, which will have instructions for how to proceed. JOSS is a little different from a traditional journal, where ideally you don't submit a single "report" but rather work your way through a checklist in an interactive/conversational way, raising any concerns you have with the package authors along the way. Once the review gets started, hopefully it will become reasonably clear, but don't hesitate to let me know if you have any questions!
