### 2024-08-06 #7080 @editorialbot
[REVIEW]: AutoBZ.jl: automatic, adaptive Brillouin zone integration of response functions using Wannier interpolation
**Submitting author:** @lxvm (<a href="http://orcid.org/0000-0003-0807-5034">Lorenzo Van Munoz</a>)
**Repository:** https://github.com/lxvm/AutoBZ.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5.4
