### 2024-08-06 #7080 @rkurchin
@lxvm note that NonlinearSolve.jl does have a `CITATION.cff` file that includes a DOI, so be sure to update that.
