+++ src/AutoBZ.jl
@@ -128,5 +128,7 @@ include("SSymmetricCompact.jl")
 export load_interp, load_autobz, load_wannier90_data
 include("wannier90io.jl")
 
+export DLRIntegral
+include("lehmann.jl")
 
 end
