+++ src/lehmann.jl
@@ -1,0 +1,5 @@
+abstract type AbstractDLRAlgorithm end
+
+struct DLRIntegral{T} <: AbstractDLRAlgorithm
+    Λ::T
+end
