### 2024-08-07 3984b68
add NonlinearSolve.jl doi
