+++ paper/paper.bib
@@ -235,5 +235,6 @@ author = {Vijay Singh and Uthpala Herath and Benny Wah and Xingyu Liao and Aldo 
   title={NonlinearSolve. jl: High-Performance and Robust Solvers for Systems of Nonlinear Equations in Julia},
   author={Pal, Avik and Holtorf, Flemming and Larsson, Axel and Loman, Torkel and Schaefer, Frank and Qu, Qingyu and Edelman, Alan and Rackauckas, Chris and others},
   journal={arXiv preprint arXiv:2403.16341},
-  year={2024}
+  year={2024},
+  doi={10.5281/zenodo.10397607}
 }
