### 2024-08-29 lxvm/AutoBZ.jl#41 @lxvm
Hi @Luthaf , thank you for the review. I can reproduce the issue and was able to successfully install the package by replacing the `version` keyword with `rev`. If this works for you I will merge the fix in #43.

This means that I expect the following to work:
```jl
using Pkg
Pkg.activate()
Pkg.add(url="https://github.com/lxvm/AutoBZ.jl.git", rev="main")
```
