### 2024-08-29 d71a5de
Create CONTRIBUTING.md (#45)
* Create CONTRIBUTING.md

New community guidelines

* Update workflow.md

add contributing section
