+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+## How to contribute to AutoBZ.jl
+
+#### **Did you find a bug?**
+
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/lxvm/AutoBZ.jl/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/lxvm/AutoBZ.jl/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
+
+#### **Did you write a patch that fixes a bug?**
+
+* Open a new GitHub pull request with the patch (read more on [pull requests](http://help.github.com/pull-requests/)).
+
+* Ensure the PR description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+#### **Do you intend to add a new feature or change an existing one?**
+
+* Suggest your change in a [GitHub discussion](https://github.com/lxvm/AutoBZ.jl/discussions/new?category=ideas) and start writing code.
+
+* Do not open an issue on GitHub until you have collected positive feedback about the change. GitHub issues are primarily intended for bug reports and fixes.
+
+* Once your proposal is positively accepted, open a new GitHub pull request with the new feature.
+
+#### **Do you have questions about the source code?**
+
+* Ask any question about how to use AutoBZ.jl in a [GitHub discussion](https://github.com/lxvm/AutoBZ.jl/discussions/new?category=q-a). Your question may also help others!
+
+* If you want to contribute to the documentation, then open a new GitHub pull request, as described above
+
+Thanks!
+
+[@lxvm](https://github.com/lxvm)
