+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"pages/app/interfaces/#Interfaces","page":"Interfaces","title":"Interfaces","text":"","category":"section"},{"location":"pages/app/interfaces/#Wannier90","page":"Interfaces","title":"Wannier90","text":"","category":"section"},{"location":"pages/app/interfaces/","page":"Interfaces","title":"Interfaces","text":"AutoBZ.load_wannier90_data\nAutoBZ.load_interp\nAutoBZ.load_autobz","category":"page"},{"location":"pages/app/interfaces/#AutoBZ.load_wannier90_data","page":"Interfaces","title":"AutoBZ.load_wannier90_data","text":"load_wannier90_data(seedname::String; bz::AbstractBZ=FBZ(), interp::AbstractWannierInterp=HamiltonianInterp, kwargs...)\n\nReturn a tuple (interp, bz) containing the requested Wannier interpolant, interp and the Brillouin zone bz to integrate over. The seedname should point to Wannier90 data to read in. Additional keywords are passed to the interpolant constructor, load_interp, while load_autobz can be referenced for Brillouin zone details. For a list of possible keywords, see subtypes(AbstractBZ) and using TypeTree; tt(AbstractWannierInterp).\n\n\n\n\n\n","category":"function"},{"location":"pages/app/interfaces/#AutoBZ.load_interp","page":"Interfaces","title":"AutoBZ.load_interp","text":"load_interp(::Type{<:HamiltonianInterp}, seed;\n    gauge=Wannier(), soc=nothing,\n    compact=:N, precision=Float64, droptol=eps(precision))\n\nLoad Hamiltonian coefficients from Wannier90 output \"seed_hr.dat\" into an AbstractHamiltonianInterp that interpolates h with unit period. The gauge keyword can be Wannier or Hamiltonian to set whether the coefficients are interpolated in the orbital basis, i.e. as is, or rotated into the band basis, i.e. the Hamiltonian eigenbasis. Additionally, the soc keyword can specify a SMatrix (from StaticArrays.jl) of twice the size of the coefficients in the seed file that is added to the Hamiltonian in orbital basis. For more details see SOCMatrix.\n\nThere are also several keywords to control the memory usage of the array: compact which may indicate the coefficient matrices are Hermitian, precision which sets the floating-point precision of the array, and droptol which skips coefficients under the given relative tolerance. Possible values of compact are:\n\n:N: do not store the coefficients in compact form\n:L: store the lower triangle of the coefficients\n:U: store the upper triangle of the coefficients\n:S: store the lower triangle of the symmetrized coefficients, (c+c')/2\n\n\n\n\n\nload_interp(::Type{<:BerryConnectionInterp}, seed;\n    coord=Lattice(), soc=nothing,\n    precision=Float64, compact=:N, droptol=eps(precision))\n\nLoad position operator coefficients from Wannier90 output \"seed_r.dat\" into a BerryConnectionInterp that interpolates (A1, A2, A3). Specify coord as Lattice or Cartesian to have the position operator interpolated in those coordinates.\n\n\n\n\n\nload_interp(::Type{<:GradientVelocityInterp}, seed, A;\n    gauge=Wannier(), vcomp=Whole(), coord=Lattice(), soc=nothing,\n    precision=Float64, compact=:N, droptol=eps(precision))\n\nLoad coefficients for a Hamiltonian and its deriv
[…]
