### 2024-08-29 4c28a54
Update workflow.md (#43)
Fix typo of `version` to `rev` in installation instructions
