### 2024-09-11 lxvm/AutoBZ.jl#46 @Luthaf
Examples fail to run
This is part of https://github.com/openjournals/joss-reviews/issues/7080

I tried running the default example from the documentation, but it failed with the following error:

```
$ julia --project kc_benchmark.jl
┌ Info: Symmetry test of the interpolant's Hamiltonian at a random k-point
│   seedname = "svo"
│   kpt = [0.9292277860683943, 0.7662948820656991, 0.7315097036118203]
│   sym = [-1.0 0.0 0.0; 0.0 0.0 -1.0; 0.0 -1.0 0.0]
└   err = 1.4540107379219425e-14
ERROR: LoadError: UndefVarError: `Ω` not defined
Stacktrace:
 [1] top-level scope
   @ ~/.../demos/kc_benchmark.jl:48
in expression starting at ~/.../demos/kc_benchmark.jl:48
```

Other examples also fail, all with `LoadError: MethodError: no method matching init_cacheval`: this is `chem_pot_test.jl`, `dos_test.jl`, and `oc_script.jl`.

------

It might make sense to run the examples on CI regularly to make sure they are always up to date with the latest version of the code.
