### 2024-09-11 lxvm/AutoBZ.jl#46 @lxvm
Hi @Luthaf, thank you for the issue. I have corrected the typo giving the first bug you mentioned in #38 and will also include the fix for the method error in https://github.com/lxvm/AutoBZCore.jl/pull/13. (The `TAI` algorithm was not fully reimplemented after breaking API changes.) I was waiting to resolve CI issues related to plotting in that release (v0.5.5) and will also include the scripts in CI as you suggested in that release. I will close the issue when v0.5.5 is merged.

Here is a quick script I use for testing that will become the CI script:
```jl
using Pkg
Pkg.activate(; temp=true)
Pkg.develop(path="") # for CI testing
# Pkg.add(url="https://github.com/lxvm/AutoBZ.jl.git", rev="v0.5.5") # for testing releases
using AutoBZ
cd(joinpath(dirname(dirname(pathof(AutoBZ))), "demos"))
include("dos_test.jl")
include("oc_test.jl")
include("chem_pot_test.jl")
include("kc_benchmark.jl")
include("oc_script.jl")
include("dos_interp.jl")
```
