### 2024-09-11 6aca58a
bump AutoBZCore compat
