+++ demos/rundemos.jl
@@ -1,0 +1,6 @@
+include("dos_test.jl")
+include("oc_test.jl")
+include("chem_pot_test.jl")
+include("kc_benchmark.jl")
+include("oc_script.jl")
+include("dos_interp.jl")
