### 2024-09-12 71c64ca
fix demos CI
