### 2024-09-12 38bf3e4
V0.5.5 (#38)
* simplify dos tutorial integer tb

* add a pythtb interface

* add bz visualization tutorials

* add load_pythtb_data

* bump v0.5.5

* fix doc

* add a note to the docs

* typo

* update pythtb docstring

* fix a pythtb gradientvelocityinterp

* bump AutoBZCore compat

* glmakie ci fixes

* add demos CI

* CI fixes

* fix demos CI
