+++ docs/src/pages/viz/band.md
@@ -1,0 +1,15 @@
+# Band diagrams
+
+Plotting band diagrams on k-paths is a common visualization tool and we show
+some examples of how to do so within `AutoBZ`. We use the
+[Brillouin.jl](https://github.com/thchr/Brillouin.jl) package to select and
+visualize k-paths.
+
+## Energy bands
+
+!!! note "In progress"
+    This will be added soon
+
+## Band velocities
+
+To overlay band velocities on the k-ath 
