### 2024-09-12 lxvm/AutoBZ.jl#46 @lxvm
I added a "Demos" CI in #38 so now all the examples are able to run.
