+++ v0.5.5/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"pages/demo/density/#Electron-density","page":"Electron density","title":"Electron density","text":"","category":"section"},{"location":"pages/demo/density/","page":"Electron density","title":"Electron density","text":"The electron density describes the number of electrons in a system. It can be calculated by integrating the DOS times a Fermi distribution over all frequencies. Often calculations of the density are needed to ensure charge self-consistency. After walking through these tutorials, continue with the demos/chem_pot_test.jl script that compares several algorithms for the calculation of the electron density of a Wannier90 Hamiltonian using the load_wannier90_data interface and a frequency-dependent self energy using the load_self_energy interface.","category":"page"},{"location":"pages/demo/density/#Model-calculation","page":"Electron density","title":"Model calculation","text":"","category":"section"},{"location":"pages/demo/density/","page":"Electron density","title":"Electron density","text":"For this tutorial and the optical conductivity tutorial we define a simple tight-binding model based on t_2g orbitals with a nearest neighbor intraband hopping and a next-nearest neighbor interband hopping.","category":"page"},{"location":"pages/demo/density/","page":"Electron density","title":"Electron density","text":"using StaticArrays\nusing OffsetArrays\nusing AutoBZ\nH = OffsetArray(zeros(SMatrix{3,3,Float64,9}, 3,3,3), -1:1, -1:1, -1:1)\n# intraband hoppings\nt = -0.25 # nearest-neighbor hopping\nH[ 1, 0, 0] = H[-1, 0, 0] =  [ 0; 0; 0;; 0; t; 0;; 0; 0; t]\nH[ 0, 1, 0] = H[ 0,-1, 0] =  [ t; 0; 0;; 0; 0; 0;; 0; 0; t]\nH[ 0, 0, 1] = H[ 0, 0,-1] =  [ t; 0; 0;; 0; t; 0;; 0; 0; 0]\n# interband hoppings\nt′ = 0.05 # next-nearest neighbor hopping\nH[ 0, 1, 1] = H[ 0,-1,-1] =  [ 0; 0; 0;; 0; 0;t′;; 0;t′; 0]\nH[ 0, 1,-1] = H[ 0,-1, 1] = -[ 0; 0; 0;; 0; 0;t′;; 0;t′; 0]\nH[ 1, 0, 1] = H[-1, 0,-1] =  [ 0; 0;t′;; 0; 0; 0;;t′; 0; 0]\nH[ 1, 0,-1] = H[-1, 0, 1] = -[ 0; 0;t′;; 0; 0; 0;;t′; 0; 0]\nH[ 1, 1, 0] = H[-1,-1, 0] =  [ 0;t′; 0;;t′; 0; 0;; 0; 0; 0]\nH[ 1,-1, 0] = H[-1, 1, 0] = -[ 0;t′; 0;;t′; 0; 0;; 0; 0; 0]\nH = HamiltonianInterp(AutoBZ.Freq2RadSeries(FourierSeries(H, period=2pi)))","category":"page"},{"location":"pages/demo/density/","page":"Electron density","title":"Electron density","text":"With this Hamiltonian we can define an ElectronDensitySolver that computes the electron density at a given temperature and scattering rate.","category":"page"},{"location":"pages/demo/density/","page":"Electron density","title":"Electron density","text":"using LinearAlgebra\nbz = load_bz(CubicSymIBZ(), Diagonal(collect(AutoBZ.period(H))))\nη = 0.1 # eV\nβ = 10.0 # 1/eV\nΣ = EtaSelfEnergy(η)\natol=1e-3\nrtol=0.0\nsol
[…]
