### 2024-09-26 #7080 @Luthaf
My only remaining concern is about the summary as explained above, I'll tick as soon as I see new proofs!
