### 2024-10-09 #7080 @rkurchin
@lxvm congrats on finishing review, we're almost done! I'll send some editorial comments shortly, and in the meantime you can start working on the other author tasks above, and send me the relevant info (version tag and DOI) once you have them. Almost there!
