### 2024-10-09 #7080 @rkurchin
Editorial remarks:
* I would move Julia citation to be after the word "Julia" rather than after "package"
* 18: need a hyphen in "low-temperature" and again on lines 20-21, 36-37, and 42 ([compound adjectives](https://www.grammarly.com/blog/parts-of-speech/compound-adjectives/))
* 19: need space before citation and comma after
* 23: hyphenate "ground-state"
* 53 "Ref. (parenthetical citation)" looks awkward, maybe just change to "algorithm of Hart et al. (cite)" instead?
* 55: change "which" to "that", also missing space before citation
* figure 1 caption: chemical formula should have subscript 3

Citations:
* make sure things are consistently sentence case (some are Title Case)
* add DOI for https://arxiv.org/abs/2406.15466
