+++ paper/paper.md
@@ -35,35 +35,35 @@ bibliography: paper.bib
 # Summary
 
 
-AutoBZ.jl is a modular Julia package [@bezansonJuliaFreshApproach2017] implementing efficient algorithms for Brillouin zone (BZ) integration, a fundamental step in electronic structure calculations of physical observables in crystals such as the density of states and the optical conductivity.
+AutoBZ.jl is a modular Julia [@bezansonJuliaFreshApproach2017] package implementing efficient algorithms for Brillouin zone (BZ) integration, a fundamental step in electronic structure calculations of physical observables in crystals such as the density of states and the optical conductivity.
 These capabilities make AutoBZ.jl an extensible framework for computational research on material properties that can compute a broad range of quantities found in experimental spectra or high-throughput screenings using Wannier interpolation and BZ integration [@mostofiWannier90ToolObtaining2008].
 Our BZ integration methods, described in Refs. [@kayeAutomaticHighorderAdaptive2023] and [@VanMunoz_et_al:2024], are high-order accurate, automatically convergent to a user-specified error tolerance, and if needed, adaptive in momentum space.
-This allows the resolution of low temperature properties of strongly interacting systems, using many-body methods such as dynamical mean-field theory[@georgesDynamicalMeanfieldTheory1996a] in which frequency-dependent electronic self-energies may produce scattering rates in the sub-meV regime. The corresponding low temperature regions of phase diagrams are typically out of reach using traditional integration algorithms, which struggle to resolve localized features in momentum space.
-AutoBZ.jl can also be used to compute ground state (i.e. $T=0$ K) properties of tight-binding models, typically derived from localized Wannier functions, with a given artificial broadening.
+This allows the resolution of low-temperature properties of strongly interacting systems, using many-body methods such as dynamical mean-field theory [@georgesDynamicalMeanfieldTheory1996a], in which frequency-dependent electronic self-energies may produce scattering rates in the sub-meV regime. The corresponding low-temperature regions of phase diagrams are typically out of reach using traditional integration algorithms, which struggle to resolve localized features in momentum space.
+AutoBZ.jl can also be used to compute ground-state (i.e. $T=0$ K) properties of tight-binding models, typically derived from localized Wannier functions, with a given artificial broadening.
 We expect it to become a widely used tool in the electronic structure community, providing accurate comparisons with experimental spectra, and a robust, automated approach for high-throughput screenings and machine learning of material properties.
 
 # Statement of need
 
 Most open-source software packages used for density functional theory plus dynamical mean-field theory, including those compatible with Wannier90 (see, e.g., Refs. [@aichhornTRIQSDFTToolsTRIQS2016; @Romero_et_al:2020; @Shinaoka_et_al:2021; @Singh_et_al:2021]), employ simple uniform grids for BZ integration, despite the fact that BZ integrands may be highly localized, e.g., near energy isosurfaces for the Green's function.
 However, these details of electronic structure may sensitively control downstream observables, so it is crucial that BZ integrals be computed in a resolved manner in material-realistic calculations. 
-In practice, this requires using dense uniform integration grids, which become compute or memory-limited in low temperature calculations involving scattering rates approaching the meV scale.
+In practice, this requires using dense uniform integration grids, which become compute or memory-limited in low-temperature calculations involving scattering rates approaching the meV scale.
 Furthermore, validating uniform integration methods requires careful convergence testing which is not always prioritized, sometimes leading to under-resolved results with spurious features.
[…]
