+++ paper/paper.bib
@@ -13,7 +13,7 @@
 }
 
 @article{Singh_et_al:2021,
-title = {{DMFTwDFT: An open-source code combining Dynamical Mean Field Theory with various density functional theory packages}},
+title = {{DMFTwDFT: An open-source code combining dynamical mean field theory with various density functional theory packages}},
 journal = {Comput. Phys. Commun.},
 volume = {261},
 pages = {107778},
@@ -39,11 +39,12 @@ author = {Vijay Singh and Uthpala Herath and Benny Wah and Xingyu Liao and Aldo 
   title={High-order and adaptive optical conductivity calculations using Wannier interpolation},
   author={{Van Mu{\~n}oz}, Lorenzo and Kaye, Jason and Barnett, Alex and Beck, Sophie},
   journal={arXiv preprint arXiv:},
+  doi = {10.48550/arXiv.2406.15466},
   year={2024}
 }
 
 @article{aichhornTRIQSDFTToolsTRIQS2016,
-  title = {{{TRIQS}}/{{DFTTools}}: {{A TRIQS}} Application for {\emph{Ab Initio}} Calculations of Correlated Materials},
+  title = {{{TRIQS}}/{{DFTTools}}: {{A TRIQS}} application for {\emph{ab Initio}} calculations of correlated materials},
   shorttitle = {{{TRIQS}}/{{DFTTools}}},
   author = {Aichhorn, Markus and Pourovskii, Leonid and Seth, Priyanka and Vildosola, Veronica and Zingl, Manuel and Peil, Oleg E. and Deng, Xiaoyu and Mravlje, Jernej and Kraberger, Gernot J. and Martins, Cyril and Ferrero, Michel and Parcollet, Olivier},
   year = {2016},
@@ -59,7 +60,7 @@ author = {Vijay Singh and Uthpala Herath and Benny Wah and Xingyu Liao and Aldo 
 }
 
 @article{bezansonJuliaFreshApproach2017,
-  title = {Julia: {{A Fresh Approach}} to {{Numerical Computing}}},
+  title = {Julia: {{A fresh approach}} to {{numerical computing}}},
   shorttitle = {Julia},
   author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
   year = {2017},
@@ -76,7 +77,7 @@ author = {Vijay Singh and Uthpala Herath and Benny Wah and Xingyu Liao and Aldo 
 }
 
 @article{georgesDynamicalMeanfieldTheory1996a,
-  title = {Dynamical Mean-Field Theory of Strongly Correlated Fermion Systems and the Limit of Infinite Dimensions},
+  title = {Dynamical mean-field theory of strongly correlated fermion systems and the limit of infinite dimensions},
   author = {Georges, Antoine and Kotliar, Gabriel and Krauth, Werner and Rozenberg, Marcelo J.},
   year = {1996},
   month = jan,
@@ -91,7 +92,7 @@ author = {Vijay Singh and Uthpala Herath and Benny Wah and Xingyu Liao and Aldo 
 }
 
 @article{hartRobustAlgorithmKpoint2019,
-  title = {A Robust Algorithm for K-Point Grid Generation and Symmetry Reduction},
+  title = {A robust algorithm for k-point grid generation and symmetry reduction},
   author = {Hart, Gus L. W. and Jorgensen, Jeremy J. and Morgan, Wiley S. and Forcade, Rodney W.},
   year = {2019},
   month = jun,
[…]
