### 2024-10-09 89bbb83
citation case fixes
