+++ paper/paper.bib
@@ -44,7 +44,7 @@ author = {Vijay Singh and Uthpala Herath and Benny Wah and Xingyu Liao and Aldo 
 }
 
 @article{aichhornTRIQSDFTToolsTRIQS2016,
-  title = {{{TRIQS}}/{{DFTTools}}: {{A TRIQS}} application for {\emph{ab Initio}} calculations of correlated materials},
+  title = {{{TRIQS}}/{{DFTTools}}: {{A TRIQS}} application for {\emph{ab initio}} calculations of correlated materials},
   shorttitle = {{{TRIQS}}/{{DFTTools}}},
   author = {Aichhorn, Markus and Pourovskii, Leonid and Seth, Priyanka and Vildosola, Veronica and Zingl, Manuel and Peil, Oleg E. and Deng, Xiaoyu and Mravlje, Jernej and Kraberger, Gernot J. and Martins, Cyril and Ferrero, Michel and Parcollet, Olivier},
   year = {2016},
@@ -156,8 +156,8 @@ author = {Vijay Singh and Uthpala Herath and Benny Wah and Xingyu Liao and Aldo 
 }
 
 @article{mostofiWannier90ToolObtaining2008,
-  title = {Wannier90: {{A}} tool for obtaining maximally-localised {{Wannier}} functions},
-  shorttitle = {Wannier90},
+  title = {{{wannier90}}: {{A}} tool for obtaining maximally-localised {{Wannier}} functions},
+  shorttitle = {wannier90},
   author = {Mostofi, Arash A. and Yates, Jonathan R. and Lee, Young-Su and Souza, Ivo and Vanderbilt, David and Marzari, Nicola},
   year = {2008},
   month = may,
