### 2024-10-10 a7c5e9f
JOSS paper (#47)
* initial commit

* fix header

* add summary

* add statement of need

* revision

* fix ref

* add ref for nonlinearsolve

* update figure

* bib typo

* tweaks

* bz tweaks

* inset caption

* Edis in summary

* Edits in Statement of need

* another edit

* iteration

* simplifications

* edits throughout

* a few edits

* few more edits

* fix gauge typo bug

* test both gauges in more cases

* add figure script+environment

* updated figure

* simplify figure

* figure tweaks

* fix refs

* another ref fix

* semicolon not comma

* final joss paper edits (#37)

Co-authore
[…]
