+++ paper/figure.jl
@@ -1,0 +1,260 @@
+using AutoBZ, HChebInterp, LinearAlgebra
+
+seed = "svo" # Wannier90 output files
+
+# model parameters
+T = 24.0 # K
+kB = 8.617333262e-5 # eV/K
+β = inv(kB*T)
+
+T₀ = 300.0 # K
+Z = 0.5
+η = (kB*pi/T₀/Z)*T^2 # eV
+Σ = EtaSelfEnergy(η)
+
+kalgs = [PTR(npt=100), PTR(npt=300), IAI()]
+kalg_labels = ["Nₖ=100", "Nₖ=300", "IAI"]
+colors = [:forestgreen, :dodgerblue, :orangered]
+
+
+# DOS calculation
+
+abstol = 1e-5
+reltol = 1e-4
+
+atol = 1e-3
+
+ω = 2.0 # eV
+μ = 12.5 # eV
+
+h, bz = load_wannier90_data(seed;
+    interp=HamiltonianInterp,
+    gauge=Wannier(),
+    bz=CubicSymIBZ(),
+)
+
+trg_interps = map(kalgs) do bzalg
+    trg_solver = TrGlocSolver(Σ, h, bz, bzalg; ω, μ, abstol, reltol)
+
+    trg_batch = let μ=μ, N=Threads.nthreads(), trg_chnl = Channel{typeof(trg_solver)}(N)
+        for _ in 1:N
+            put!(trg_chnl, TrGlocSolver(Σ, h, bz, bzalg; ω, μ, abstol, reltol))
+        end
+        BatchFunction() do ωs
+            out = Vector{ComplexF64}(undef, length(ωs))
+            Threads.@threads for n in 1:N
+                _solver = take!(trg_chnl)
+                for i in n:N:length(out)
+                    AutoBZ.update_trgloc!(_solver; ω = ωs[i], μ)
+                    sol = solve!(_solver)
+                    out[i] = sol.value
+                end
+                put!(trg_chnl, _solver)
+            end
+            return out
+        end
+    end
+    hchebinterp(trg_batch, -ω, ω; atol)
+end
+
+#=
+# electron density calculation
+
+falg = QuadGKJL()
+abstol = 1e-5
+reltol = 1e-4
+
+atol = 1e-3
+
+μ = 12.5 # eV
+μmax = 2.0 # eV
+
+h, bz = load_wannier90_data(seed;
+    interp=HamiltonianInterp,
+    gauge=Hamiltonian(),
+    bz=CubicSymIBZ(),
+)
+
+ρ_interps = map(kalgs) do bzalg
+    ρ_solver = ElectronDensitySolver(h, bz, bzalg, Σ, falg; β, μ, abstol, reltol)
+
+    ρ_batch = let β=β, N=Threads.nthreads(), ρ_chnl = Channel{typeof(ρ_solver)}(N)
+        for _ in 1:N
+            put!(ρ_chnl, ElectronDensitySolver(h, bz, bzalg, Σ, falg; β, μ, abstol, reltol))
+        end
+        BatchFunction() do μs
+            out = Vector{Float64}(undef, length(μs))
+            Threads.@threads for n in 1:N
+                _solver = take!(ρ_chnl)
+                for i in n:N:length(out)
+                    AutoBZ.update_density!(_solver; β, μ=μs[i])
+                    sol = solve!(_solver)
+                    out[i] = sol.value
+                end
+                put!(ρ_chnl, _solver)
+            end
+            return out
+        end
+    end
+    hchebinterp(ρ_batch, μ-μmax, μ+μmax; atol)
+end
+=#
+
+
[…]
