### 2024-10-14 #7080 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
