### 2024-10-20 #7080 @Kevin-Mattheus-Moerman
@lxvm as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

- [ ] Software license tag listed here matches a tagged release

- [ ]
