+++ CITATION.cff
@@ -1,0 +1,42 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Munoz
+  given-names: Lorenzo
+  name-particle: Van
+  orcid: "https://orcid.org/0000-0003-0807-5034"
+- family-names: Beck
+  given-names: Sophie
+  orcid: "https://orcid.org/0000-0002-9336-6065"
+- family-names: Kaye
+  given-names: Jason
+  orcid: "https://orcid.org/0000-0001-8045-6179"
+doi: 10.5281/zenodo.13915617
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Munoz
+    given-names: Lorenzo
+    name-particle: Van
+    orcid: "https://orcid.org/0000-0003-0807-5034"
+  - family-names: Beck
+    given-names: Sophie
+    orcid: "https://orcid.org/0000-0002-9336-6065"
+  - family-names: Kaye
+    given-names: Jason
+    orcid: "https://orcid.org/0000-0001-8045-6179"
+  date-published: 2024-10-20
+  doi: 10.21105/joss.07080
+  issn: 2475-9066
+  issue: 102
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 7080
+  title: "AutoBZ.jl: Automatic, adaptive Brillouin zone integration
+    using Wannier interpolation"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.07080"
+  volume: 9
+title: "AutoBZ.jl: Automatic, adaptive Brillouin zone integration using
+  Wannier interpolation"
