### 2024-10-20 #7080 @Kevin-Mattheus-Moerman
@lxvm On the name. I'll ping @openjournals/dev to see if they know a solution. Once that is done I will close this issue.

@openjournals/dev The name needs to render as `Lorenzo Van Munoz`, can you help with that? Thanks
