### 2024-10-22 #7080 @xuanxu
Info [updated](https://joss.theoj.org/papers/10.21105/joss.07080) to `Lorenzo Van Munoz`
