### 2024-10-29 62dbae3
add pythtb band visualization example
