### 2024-10-31 #7080 @Kevin-Mattheus-Moerman
@openjournals/dev @xuanxu do I have to reaccept? Or can I close this now? Thanks
